feat: add Speakers component and route, update Layout for navigation, and remove SpeakersManagementTab
